reference/netology_pd_diplom/requirements.txt
django~=5.0
djangorestframework~=3.14.0
djangorestframework-yaml~=2.0.0
django-rest-passwordreset>=1.3.0
celery~=5.4.0
requests~=2.31.0
ujson~=5.9.0
pyyaml~=6.0.0
python-dotenv~=1.0.1
psycopg2-binary
flower~=2.0.1
redis~=5.0.4
pytest~=8.2.0
pytest-cov~=5.0.0
pytest-django~=4.8.0
model-bakery~=1.18.0